Modify restore command to restore from a specified Commit
This adds a new argument to the `restore` command to allow for restoring
a full back up or a specified file from a specific back up revision. The
argument can either be the value `HEAD`, which performs a restore from
the most recent backup and was the old default, or it can be a blobKey
representing a Commit.

The argument is referred to as a Commit BlobKey and can be used from the
output provided by `compareAll`. However, it's not clear if this is the
*ideal* reference to a particular Commit.

The intention is to restore a specific file that has been detected by
`compareAll` to have changed some time during the back up history. By
using the blobKey of a Commit output by compareAll and the path of a
changed file in that same output, the `restore` command can pull the
version of the file from the backup set before it had bit rot.